+++ src/routes/tasks/components/AggregateTaskSiderContent.tsx
@@ -2,6 +2,7 @@ import { Fragment } from 'react';
 
 import AlgoSiderSection from './AlgoSiderSection';
 import InModelsSiderSection from './InModelsSiderSection';
+import ModelSiderSection from './ModelSiderSection';
 import TimelineSiderSection from './TimelineSiderSection';
 
 import { AggregatetupleT } from '@/modules/tasks/TuplesTypes';
@@ -19,6 +20,11 @@ const AggregateTaskSiderContent = ({
         <Fragment>
             <AlgoSiderSection task={task} />
             <InModelsSiderSection task={task} />
+            <ModelSiderSection title="Out model key" model={task.out_model} />
+            <PermissionSiderSection
+                title="Out model permissions"
+                permission={task.permissions.process}
+            />
             <PermissionSiderSection permission={task.permissions.process} />
             <TimelineSiderSection />
         </Fragment>