+++ src/routes/tasks/components/TrainTaskSiderContent.tsx
@@ -3,6 +3,7 @@ import { Fragment } from 'react';
 import AlgoSiderSection from './AlgoSiderSection';
 import DatasetSiderSection from './DatasetSiderSection';
 import InModelsSiderSection from './InModelsSiderSection';
+import ModelSiderSection from './ModelSiderSection';
 import TimelineSiderSection from './TimelineSiderSection';
 
 import { TraintupleT } from '@/modules/tasks/TuplesTypes';
@@ -21,6 +22,11 @@ const TrainTaskSiderContent = ({
             <AlgoSiderSection task={task} />
             <DatasetSiderSection task={task} />
             <InModelsSiderSection task={task} />
+            <ModelSiderSection title="Out model key" model={task.out_model} />
+            <PermissionSiderSection
+                title="Out model permissions"
+                permission={task.permissions.process}
+            />
             <PermissionSiderSection permission={task.permissions.process} />
             <TimelineSiderSection />
         </Fragment>