+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 -   Display channel name in header instead of node ID (#97)
+-   Add out_model key and permissions in Sider (#98)
 
 ## [0.1.0] - 2021-08-04